Refactor: Move station logic to Vehicle.Station.cpp
This commit departs Synchronised departure, loading/waiting, departing, travelling, and arriving functions. They're now are now stationed in Vehicle.Station.cpp
