+++ src/openrct2/libopenrct2.vcxproj
@@ -1126,6 +1126,7 @@
     <ClCompile Include="ride\Vehicle.cpp" />
     <ClCompile Include="ride\Vehicle.Animation.cpp" />
     <ClCompile Include="ride\Vehicle.Sound.cpp" />
+    <ClCompile Include="ride\Vehicle.Station.cpp" />
     <ClCompile Include="ride\VehicleData.cpp" />
     <ClCompile Include="ride\VehicleGeometry.cpp" />
     <ClCompile Include="ride\VehicleRiderControl.cpp" />
